refactor(kernel): extract Recycler[T] into internal/kernel/recycler
Move the generic typed object pool (Recycler[T] + objectBucket + NewRecycler)
out of internal/kernel/buffer into a new internal/kernel/recycler package,
keeping the interface and nil-factory -> nil behaviour verbatim. Pure move,
no behaviour change. Re-point the two consumers + their tests:
  - internal/service/logger/pool.go (recordPool)
  - internal/service/logger/middleware/async (sink pool + internal tests)

recycler ships CLAUDE.md + README.md, the kernel_consumers visibility +
audit_srcs filegroup, and is wired into //:audit_sources.

Bundled toolchain-upgrade fallout (the devcontainer's ktn-linter was a
regressed v1.3.206 emitting ~200 false positives; /ktn upgraded it to
v1.34.0, and gofumpt is now v0.10.0 — both stricter than the versions that
last baselined main, and neither runs in CI):
  - converge two pre-existing gofumpt drifts (pkg/v1 test files) to the
    v0.10.0 fixpoint
  - relocate the pre-existing BenchmarkBuild_ZeroAlloc into
    builder_bench_test.go (KTN-TEST-SUFFIX) + ship internal/service/logger/BENCH.md
  - drop the carried BenchmarkRecyclerSteadyState (superseded by the planned
    zero-alloc invariant gate)

Part of the kernel-recycler consolidation (ADR 0010, #30).
