+++ internal/kernel/recycler/CLAUDE.md
@@ -1,0 +1,47 @@
+# internal/kernel/recycler/
+
+## Purpose
+
+The SDK's generic object-recycling primitive. Stdlib-only, domain-neutral.
+`Recycler[T]` wraps a `sync.Pool` to recycle ANY pointer-sized typed object
+(event records, attr slices, scratch structs) without the boxing cost on
+Get/Put. The byte-slice pool (`internal/kernel/buffer`) and the codec scratch
+buffer pool (`internal/core/codec/scratch`) are built ON this primitive — the
+mechanism lives here, the capacity thresholds stay with the consumers
+(ADR 0010).
+
+## Contents
+
+| Primitive | Surface | Use case |
+|---|---|---|
+| `Recycler[T]` | `NewRecycler[T any](newFn func() T)` → `Get() T` / `Put(v T)` | recycle typed objects; no reset |
+| `CappedRecycler[T]` *(commit 3)* | `NewCappedRecycler[T any](newFn, resetFn, capOfFn, maxCap)` → `Get`/`Put` | adds reset-on-Put + cap-discard |
+
+## Conventions
+
+- **Recycler[T] has no reset.** Consumers that need cleanup either reset
+  before Put (logger `Builder.Send`, async `data[:0]`) or use
+  `CappedRecycler[T]`.
+- **CappedRecycler[T] is reset-on-Put + discard-before-reset.** Over-cap
+  values are orphaned (never reset, never repooled) — this preserves the
+  codec scratch detach contract. Reset is NOT a memory wipe.
+- **Generic recyclers do NOT nil-check.** Public/package-level wrappers
+  (`buffer.Put`, `scratch.ReleaseBuffer`) keep nil guards where nil is part
+  of their accepted API — a generic `T any` cannot nil-check.
+- **Pointer-like T only.** `sync.Pool` boxes everything via `any`; a
+  non-pointer-sized T pays an allocation on every Put.
+
+## Do NOT
+
+- Keep a reference to a value after `Put`.
+- Use `Recycler[T]` for value types larger than a few words.
+- Add worker pools, queues, metrics, policy objects, or lifecycle managers
+  here — this is a primitive, not a framework (ADR 0010).
+
+## Verification
+
+```
+bazel test --config=race //internal/kernel/recycler:recycler_test
+# zero-alloc gate runs HORS race (race instrumentation perturbs allocs):
+bazel test --config=pure //internal/kernel/recycler:recycler_test
+```
