+++ internal/kernel/recycler/recycler_internal_test.go
@@ -1,4 +1,4 @@
-package buffer
+package recycler
 
 import "testing"
