+++ internal/service/logger/middleware/async/BUILD.bazel
@@ -21,8 +21,8 @@ go_library(
     ],
     deps = [
         "//internal/core/logger",
-        "//internal/kernel/buffer",
         "//internal/kernel/errs",
+        "//internal/kernel/recycler",
         "//internal/kernel/ring",
     ],
 )
@@ -49,7 +49,7 @@ go_test(
     deps = [
         "//internal/core/logger",
         "//internal/core/logger/level",
-        "//internal/kernel/buffer",
         "//internal/kernel/errs",
+        "//internal/kernel/recycler",
     ],
 )
