+++ internal/service/logger/middleware/async/async_sink.go
@@ -18,8 +18,8 @@ import (
 	"sync"
 
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
-	"github.com/kitsunium/sdk/internal/kernel/buffer"
 	"github.com/kitsunium/sdk/internal/kernel/errs"
+	"github.com/kitsunium/sdk/internal/kernel/recycler"
 	"github.com/kitsunium/sdk/internal/kernel/ring"
 )
 
@@ -35,7 +35,7 @@ type asyncSink struct {
 	// queue is the lock-free ring buffer holding pending entries.
 	queue ring.Queue[*recordEntry]
 	// pool recycles *recordEntry values to avoid per-write allocation.
-	pool buffer.Recycler[*recordEntry]
+	pool recycler.Recycler[*recordEntry]
 	// policy selects the saturation behaviour at Write time.
 	policy DropPolicy
 	// onDrop fires for every entry discarded by the policy; never nil.
@@ -107,7 +107,7 @@ func New(downstream corelogger.Sink, cfg Config) corelogger.Sink {
 		errCallback = noopOnError
 	}
 	//: recycler keeps per-entry allocation off the hot path.
-	pool := buffer.NewRecycler[*recordEntry](newRecordEntry)
+	pool := recycler.NewRecycler[*recordEntry](newRecordEntry)
 	//: build the sink with channels primed for the drainer lifecycle.
 	out := &asyncSink{
 		downstream:  downstream,
